+++ taskmanager.h
@@ -5,11 +5,17 @@
 #include <thread>
 #include <functional>
 #include <vector>
+#include <list>
 #include <queue>
 #include <string>
 #include <atomic>
+#include <iostream>
+#include <array>
+
+#include "loggable.h"
 
 class TaskManager {
+    typedef std::pair<bool, std::list<Loggable::Message>> JobResult;
 public:
     TaskManager();
     ~TaskManager();
@@ -20,18 +26,20 @@ public:
     bool busy() const;
     void wait();
     void printProgress() const;
+    void printProgress(const JobResult& result, const std::string& source, const std::string& destination) const;
 
 private:
     void loop();
     bool loopCondition() const;
     bool waitCondition() const;
-    static bool job(const std::string& source, const std::string& destination);
-
+    static JobResult job(const std::string& source, const std::string& destination);
+    static void printLog(const JobResult& result, const std::string& source, const std::string& destination);
 private:
     std::atomic<uint32_t> busyThreads;
     std::atomic<uint32_t> maxTasks;
     std::atomic<uint32_t> completeTasks;
     bool terminate;
+    mutable std::mutex printMutex;
     mutable std::mutex queueMutex;
     std::mutex busyMutex;
     std::condition_variable loopConditional;