+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
+## MLC 1.1.0 (July 23, 2023)
+- New logging system
+- Artist, Album artist, Album and Title are now utf16 encoded, should fix broten titles
+- BPM tag is now rounded, as it supposed to be by spec
+- Lyrics is not set now for USLT tag for unsychronized lyrics is not supported in LAME
+- Date bug fix, it was MMDD instead of standard DDMM
+
 ## MLC 1.0.1 (July 21, 2023)
-### Added multithreaded encoding
-### Only the first artist vorbis tag is left in the id3 tags
+- Added multithreaded encoding
+- Only the first artist vorbis tag is left in the id3 tags
 
 ## MLC 1.0.0 (July 19, 2023)
-### Initial release, only core functionality
+- Initial release, only core functionality